+++ feature/feature-home/src/main/AndroidManifest.xml
