+++ feature/feature-home/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+</resources>
