+++ feature/feature-mypage/src/main/AndroidManifest.xml
