+++ feature/feature-mypage/src/main/java/com/ggne/feature_mypage/MypageFragment.kt
@@ -1,0 +1,18 @@
+package com.ggne.feature_mypage
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.ggne.feature_more.R
+
+class MypageFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_mypage, container, false)
+    }
+}
