+++ feature/feature-mypage/src/main/res/navigation/nav_mypage.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_mypage"
+    app:startDestination="@id/fragment_mypage">
+
+    <fragment
+        android:id="@+id/fragment_mypage"
+        android:name="com.ggne.feature_mypage.MypageFragment"
+        android:label="마이페이지"
+        tools:layout="@layout/fragment_mypage" />
+
+</navigation>
