+++ feature/feature-subway/src/main/java/com/ggne/feature_subway/SubwaySearchFragment.kt
@@ -1,0 +1,16 @@
+package com.ggne.feature_subway
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+class SubwaySearchFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_subway_search, container, false)
+    }
+}
