+++ feature/feature-subway/src/main/res/navigation/nav_subway.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_subway"
+    app:startDestination="@id/fragment_subway_search">
+
+    <fragment
+        android:id="@+id/fragment_subway_search"
+        android:name="com.ggne.feature_subway.SubwaySearchFragment"
+        android:label="지하철 검색"
+        tools:layout="@layout/fragment_subway_search" />
+
+</navigation>
