+++ feature/feature-teatime/src/main/java/com/ggne/feature_teatime/TeatimeFragment.kt
@@ -1,0 +1,17 @@
+package com.ggne.feature_teatime
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class TeatimeFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_teatime, container, false)
+    }
+}
