+++ feature/feature-teatime/src/main/res/navigation/nav_teatime.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_teatime"
+    app:startDestination="@id/fragment_teatime">
+
+    <fragment
+        android:id="@+id/fragment_teatime"
+        android:name="com.ggne.feature_teatime.TeatimeFragment"
+        android:label="티타임"
+        tools:layout="@layout/fragment_teatime" />
+
+</navigation>
